Initial fork with modifications.
This was forked from tamersaadeh/oab-java which is a fork of flexiondotorg/oab-java6. It was modified to suit the needs of Team DIANA and is highly oriented to our specific needs. However, it can still be of use to others.

Our goal is to create a deb repository with packages we may need to be up to date that aren't provided through other means.

The original commits were squached into this as they provide no added value to the current code (other than being a seed). However, below is the commit messages of the original history of the code, kept for historical reasons.

This is a combination of 67 commits.

The first commit's message is:

Initial commit

(commit numbering error, add one to the value reported below)

The first commit's message is:

Minor formatting improvements

This is the 2nd commit message:

Futher testing (maverick) showed that 'build-essential' is also required.

This is the 3rd commit message:

Updated documentation

This is the 4th commit message:

Minor documentation format fix

This is the 5th commit message:

Updated documentation.

This is the 6th commit message:

Updated to use dynamic version detection throughout. Fixed package installation when upgrading. Minor documentation updates.

This is the 7th commit message:

Minor documentation update

This is the 8th commit message:

Typo corrected

This is the 9th commit message:

Minor documentation edit

This is the 10th commit message:

Added install the required Java dependencies. Fixed install of 'ia32-sun-java6-bin' on 64-bit systems.

This is the 11th commit message:

Updated build requirements

This is the 12th commit message:

Fixed Java browser plugin install when a supported browser is not installed

This is the 13th commit message:

Browser plugin install, needs more thought.

This is the 14th commit message:

Fixed install of Java browser plugin

This is the 15th commit message:

Minor documentation updates

This is the 16th commit message:

Silence 'update-alternatives'

This is the 17th commit message:

Minor corrections

This is the 18th commit message:

Migrated to a solution that create a local 'apt' repository

This is the 19th commit message:

Closes issue #1. Closes issue #2.

This is the 20th commit message:

Migrated to new common functions

This is the 21st commit message:

Added GnuPG signing of the local 'apt' repository.

This is the 22nd commit message:

Ooops! Commented in the build.

This is the 23rd commit message:

Documentation updated.

This is the 24th commit message:

Added support for Java6u31 when the upstream script tags it as stable. Other fixes and improveents, closes #3, closes #4, closes #7 and closes #8.

This is the 25th commit message:

For rraptorr/sun-java6.git, the script was using git protocol instead of https. This leads to problems behind some firewalls.

This is the 26th commit message:

Fixed GPG key creation on VMware ESX Server, closes #11. Fixed clone of the 'sun-java6' repository for users behind restrictive firewalls.

This is the 27th commit message:

Added dynamic determination of Java package URLs and sizes. Added an option (-c) to optionally clean .deb package, closes #10

This is the 28th commit message:

Minor documentation edits

This is the 29th commit message:

Updated script to build Oracle Java 7 instead.

This is the 30th commit message:

Huge refactoring of oab-java7.sh. Split to many smaller scripts so that managing it will become much easier. Also, now it should package both sun-java6 and oracle-java7. Still broken, I just need to replace the utilies used by the common.sh or figure out how to implement it everywhere.

This is the 31st commit message:

renamed the script name since it builds both java 6 and 7

This is the 32nd commit message:

added common.sh to every script that uses it

This is the 33rd commit message:

changed sh to bash to fix bugs and other fixes

This is the 34th commit message:

fixed another issue with common.sh

This is the 35th commit message:

fixed a huge typo and updated script access again

This is the 36th commit message:

minor docs update

This is the 37th commit message:

fixed perms

This is the 38th commit message:

more fixes, updated README, CHANGES, and LICENSE.

This is the 39th commit message:

updated documentation

This is the 40th commit message:

precise is still not supported!

This is the 41st commit message:

many bug fixes and typos. Also, a small refatoring of variables

This is the 42nd commit message:

added the common.sh into the scripts directory so I can modify the parts I don't need and to make it simpler, idlely I would like to get rid of it entirely!

This is the 43rd commit message:

escaped many inputs to make sure that it won't crash, the only probalem is that the script does not exit on error

This is the 44th commit message:

added the issue to the TODO list

This is the 45th commit message:

Updated scripts to build on Precise (see https://github.com/rraptorr/sun-java6/issues/5#issuecomment-4312028)

This is the 46th commit message:

check if precise

This is the 47th commit message:

more bug fixes, now builds packages correctly, but the releases file is still faultly according to apt

This is the 48th commit message:

Created release file with apt-ftparchive.

This is the 49th commit message:

Merge branches 'release-file-fix' and 'master'

This is the 50th commit message:

removed many parts of common.sh that are not being used and simplified the code

This is the 51st commit message:

Huge documentation refactoring and some other fixes.

This is the 52nd commit message:

minor documentation updates

This is the 53rd commit message:

typo

This is the 54th commit message:

created a workaround that fixes that causes .deb files to be found

This is the 55th commit message:

This prevents the Filename field from containing a path other than './'.
Having a filename with a path was generating the following error
when running 'apt-get install sun-java6-jre':

Failed to fetch file:///var/local/oab/deb//var/local/oab/deb//sun-java6-jre_6.31-1~precise1_all.deb  File not found

The path was appearing twice and eliminating it from the filename field
addressed the issue and allowed a successful install.

Thanks to Jeff Cooper <jeff.cooper@logicpd.com> for fixing this.

This is the 56th commit message:

minor changes

This is the 57th commit message:

This should simplify the parsing of the download url for java 6

This is the 58th commit message:

This should fix the parse of the correct url for java 7

This is the 59th commit message:

make it similar to the java6 fix

This is the 60th commit message:

Removed the work around for precise, since it was fixed.

See:
- https://github.com/rraptorr/sun-java6/issues/5
- https://bugs.launchpad.net/ubuntu/+source/ia32-libs/+bug/947495

This is the 61st commit message:

Updated URL parsing

This is the 62nd commit message:

Fixed download links - Oracle is expecting a cookie with the request

This is the 63rd commit message:

fix url for java 7

This is the 64th commit message:

drop java 6

This is the 65th commit message:

Update README.md

This is the 66th commit message:

initial bring up to 2015 and usage for team diana